Clean code/folder structure
This commit moves the express endpoints to the /routes folder.
This makes the app.js cleaner and make it more modular for creating/moving new endpoints.
This commit also modifies the eslint configuration file for the project.
