+++ .eslintrc.js
@@ -1,10 +1,12 @@
 module.exports = {
   "env": {
     "es6": true,
-    "node": true
+    "node": true,
+    "mocha": true
   },
   "extends": "eslint:recommended",
   "parserOptions": {
+    "ecmaVersion": 2017,
     "sourceType": "module"
   },
   "rules": {
@@ -16,16 +18,24 @@ module.exports = {
       "error",
       "unix"
     ],
-    "quotes": [
-      "error",
-      "double"
-    ],
     "semi": [
       "error",
       "always"
     ],
+    "no-console" : "off",
     "guard-for-in": 0,
     "prefer-promise-reject-errors": 2,
     "no-invalid-this": 0,
-  }
+  },
+  "globals": {
+    "jasmine": true
+  },
+  "overrides": [
+    {
+        "files": "containers/map-api/test/*.js",
+        "rules": {
+            "no-unused-vars": "off"
+        }
+    }
+]
 };
