--- containers/map-api/public/boothList.html
@@ -1,9 +1,0 @@
-<html>
-  <head>
-    <title>Main Page</title>
-  </head>
-  <body>
-    <a href="/">Home</a> / <a href="/booth">Add Booth</a> / <a href="/beacon">Add Beacon</a> / <a href="/event">Add Event</a> <br />
-    Hello, World!
-  </body>
-</html>
