+++ containers/map-api/public/index.html
@@ -3,8 +3,22 @@
     <title>Main Page</title>
   </head>
   <body>
-    <a href="/">Home</a> / <a href="/booth">Add Booth</a> / <a href="/beacon">Add Beacon</a> / <a href="/event">Add Event</a>
-    / <a href="/booth/list">Booth List</a> / <a href="/beacon/list">Beacon List</a> / <a href="/event/list">Event List</a><br />
-    Hello, World!
+    <p>Hello, World!</p>
+    <p>Endpoints</p>
+    <p>GET</p>
+    <ul>
+      <li>/beacons</li>
+      <li>/booths</li>
+      <li>/events</li>
+      <li>/beacons/:beaconId</li>
+      <li>/booths/:boothId</li>
+      <li>/events/:eventId</li>
+    </ul>
+    <p>POST</p>
+    <ul>
+      <li>/beacons/add</li>
+      <li>/booths/add</li>
+      <li>/events/add</li>
+    </ul>
   </body>
 </html>
